requirements.txt
PyMySQL==0.9.3
pandas==0.25.3
requests==2.22.0
selenium==3.141.0
PyYAML==5.1.2
pymongo==3.7.2

